Review logging setup and configuration
Just run GHIB_VERBOSE=1 jenkins-ghb to have debug log.
